rename RemoteType to RemoteResolver
