+++ remote-graph/Makefile
@@ -8,6 +8,7 @@ unit-test:
 	$$(npm bin)/mocha -r ts-node/register RemoteType.test.ts
 
 integration-test:
-	cd integration-tests && $$(npm bin)/mocha -r ts-node/register integration.test.ts
+	cd integration-tests && \
+	$$(npm bin)/mocha -r ts-node/register integration.test.ts --timeout 10000
 
 test: unit-test integration-test
